refactor: location of agents
